Revert "refactor(web): logout を features/sessions に統合しリソース名へ揃える"
This reverts commit cf9cfdc9fa54ef5f3b39f4352e5df67e49bc9747.
