+++ application/apps/web/src/client/components/ui/layout/nav-sheet/index.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/client/components/shadcn/sheet'
 import NavMenu from '@/client/components/ui/navigation/nav-menu'
 import { getVisibleMenuItems } from '@/client/entities/navigation'
-import { LogoutButton } from '@/client/features/sessions'
+import { LogoutButton } from '@/client/features/logout'
 
 interface Props {
   isLoggedIn: boolean
