+++ application/apps/web/src/client/features/logout/sidebar-logout-button.tsx
